third_party/blink/web_tests/fast/css/gradient-unsupported-color-spaces-expected.txt

CONSOLE WARNING: display-p3 is not yet supported as a gradient interpolation space.
CONSOLE WARNING: a98-rgb is not yet supported as a gradient interpolation space.
CONSOLE WARNING: rec2020 is not yet supported as a gradient interpolation space.
CONSOLE WARNING: prophoto-rgb is not yet supported as a gradient interpolation space.
CONSOLE WARNING: display-p3 is not yet supported as a gradient interpolation space.
CONSOLE WARNING: a98-rgb is not yet supported as a gradient interpolation space.
CONSOLE WARNING: rec2020 is not yet supported as a gradient interpolation space.
CONSOLE WARNING: prophoto-rgb is not yet supported as a gradient interpolation space.
CONSOLE WARNING: display-p3 is not yet supported as a gradient interpolation space.
CONSOLE WARNING: a98-rgb is not yet supported as a gradient interpolation space.
CONSOLE WARNING: rec2020 is not yet supported as a gradient interpolation space.
CONSOLE WARNING: prophoto-rgb is not yet supported as a gradient interpolation space.
